+++ src/convert_roms/test_rvz_to_sevenzip_iso.rs
@@ -1,7 +1,7 @@
-use super::super::database::*;
 use super::super::import_dats;
 use super::super::import_roms;
 use super::*;
+use crate::sevenzip::ArchiveCompression;
 use std::path::{Path, PathBuf};
 use tempfile::{NamedTempFile, TempDir};
 use tokio::fs;
@@ -71,7 +71,7 @@ async fn test() {
         false,
         false,
         true,
-        &None,
+        &ArchiveCompression::Default,
         false,
     )
     .await