+++ src/convert_roms/test_sevenzip_to_sevenzip.rs
@@ -1,7 +1,7 @@
-use super::super::database::*;
 use super::super::import_dats;
 use super::super::import_roms;
 use super::*;
+use crate::sevenzip::ArchiveCompression;
 use std::path::{Path, PathBuf};
 use tempfile::{NamedTempFile, TempDir};
 use tokio::fs;
@@ -75,7 +75,7 @@ async fn test() {
         true,
         false,
         true,
-        &None,
+        &ArchiveCompression::Default,
         false,
     )
     .await